MFC 1.41: kern.ps_arg_max_cache is a long, not an int.  This breaks ps(1) without /proc on 4.x on alpha.

+++ lib/libkvm/kvm_proc.c
@@ -670,7 +670,7 @@ kvm_getargv(kd, kp, nchr)
 	int oid[4];
 	int i;
 	size_t bufsz;
-	static int buflen;
+	static unsigned long buflen;
 	static char *buf, *p;
 	static char **bufp;
 	static int argc;
